add kube-router.io/bgp.max.paths annotation
Add the ability to define a max number of paths out of the cluster.
This is useful for debugging on-premises environments where upstream
routers may have issues with asynchronous routing (due to ECMP VIP).

The state of which nodes have selected themselves to be in the available
paths is available in the kube-router.io/bgp.state.paths annotation.
You should be able to steer which nodes are the L4 director by setting
this as well.

At this time, there is no controller watching for nodes to be deleted or
become NotReady and update the annotation.