feat(Makefile): make local builds behave like CI
RUNTIME_BASE and BUILDTIME_BASE are both consumed in the Dockerfile and
handle setting the base containers used for both building and running
kube-router. These are handled differently in CI using GitHub actions
than they are for local builds.

This updates the Makefile so that local builds behave the same way that
CI based builds in GitHub are when it comes to these bases.