fix(dsr): change grpc resolver to passthrough
When we used DialContext() before it was deprecated, it manually set the
default resolver to passthrough. NewClient() appears to not do this, and
then several versions of grpc ago, they defaulted to the DNS resolver
which broke our DSR implementation with CRI compatible container
engines.

I'm not 100% sure that there isn't a better way to specify this, but I
spent 20 - 30 minutes poking around the gRPC code base and I don't think
that I can see an easy way to do this outside of specifying it this way.

Furthermore, the project itself seems to advocate for this approach in
comments like those on: https://github.com/grpc/grpc-go/issues/1846