feat: add health check to host route sync
Host route syncing is an important part of kube-router that is run in
its own goroutine, so we should be checking it for health the way we do
the major controllers to make sure that it never stops functioning.

Additionally, as the health check controller has been continuously added
on to, we also make some modifications here to make it a bit more robust
and scalable.