feat(linux_routes.go): add a boolean to InjectRoute()
Adding a boolean to the return values of InjectRoute() allows us to
track when a route is actually inserted vs. when it was called. This
helps us better represent HostRoutesStaleAddedCounter metric.

Without this, the administrator will see a bunch of routes seemingly
being added every time we call checkCacheAgainstBGP() when, in fact,
most of them are tossed out depending on the user's mesh subnet
settings.